RUN command needs to be turned into an untokenized script and re-tokenized.
If a list is directly passed to the interpreter, none of transformations that
happen during tokenization will occur.  E.g.:

run [ sum 5 2 ]

The list argument to run will initially be tokenized as a list of quoted words.
Importantly, the numbers 5 and 2 will not be tokenized as numbers.

Before the list is passed to the interpreter, it should be flattened into a
textual script and tokenized without the outer square braces.  This ensures
that the inner commands are tokenized as WORD NUMBER NUMBER.