Added escape for comment char (\; -> ;).
It's worth it to note that some of the parsley behavior confused me for a long
time, here.  Essentially, the `<...>` shortcut for taking a slice of the token
stream disregards any return values that were produced by child productions.

That means if you have an escape character terminal production that returns
the escaped character and drops the leading backslash, the production that
uses the terminal will revert to whatever was in the original token stream
if you use `<...>` to collect some list data.  You need to use the more clunky:

  list_rule+:bound_var -> ''.join(bound_var)

syntax.  Also worth noting is that if this throws errors about receiving a bool
value when a string was expected, it means that in one of your child productions
you are returning a match (True) instead of the parse for that production.