Added REDUCE, FLOAT, DEC.STR; updated INT.
UCBLogo allows manipulating numbers as strings in some contexts.  E.g. `last 35`
evaluates to `5`.  I consider this implicit conversion confusing.  E.g.:

  show reduce [sum last ?1 last ?2] [[a 10] [b 20] [c 30]]

is a toy example I was working up to show how to use reduce with lists.  I was
quite surprised that UCBLogo evaluated this as `10`!  I finally realized that
reduce performed the following operations:

  20 + 30  ;last elements of [b 20] and [c 30]
  10 + 0   ;last elements of [a 10] and 50

In logopy, numbers and strings can't be converted implicitly.  Instead, I've
changed the behavior of INT and FLOAT and added a DEC.STR command.  INT and FLOAT
will attempt to convert a string to an integer or float.  Both commands assume the
string is a base 10 representation of the number.  The DEC.STR command converts
a number to its string base-10 representation.