Coordinate mapping issues - first attempt
I discovered that my angle correction in an earlier change was actually
symptomatic of a larger issue.  Namely, the coodinate system for SVG
starts with the origin in the upper left, x increases to the right, and
y increases going down.

In the Logo coordinate system, the origin is in the center of the screen.
X increases to the right, and Y increases going up.

Additionally, a heading of 0 degrees in the Logo system is 90 degrees
in a typical Cartesian system.

Needless to say, some drawings in one backend don't resemble drawings
in the other.

I've attempted to translate the coordinates by mentally rotating the
SVG system 90 degrees anti-clockwise.  This makes the origin in the lower
left, and X increases up and Y increases to the right.
I switch the x and y coordinates when plotting SVG lines, so now x is on
the Y-axis and y is on the X-axis.  Then I apply a rotate transformation
of -90 degrees to the line about the origin.

This seems to be working for the colorsquare-- the rotation is in the
correct direction, the colors are in the correct order.