Added EXT.UNFILLED command.
While the Python turtle module does a pretty good job not filling in holes
in any drawings, my SVG transltions need some help.

The EXT.UNFILLED extension tells the SVG backend to create masks for the
holes in filled components of the SVG output.  For the Tk backend, it
is it simply runs the instructions passed to it without any special
handling.